requirements.txt
-e .
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
matplotlib==3.1.0
notebook==5.7.8
numpy==1.16.4
Pillow==6.2.0